Update Dependencies to Jersey 1.9
Simplify the proxy class name to remove the protocol and the port number so that it only includes the hostname and the path elements. This makes the class name a little bit easier to read and use.

Added accessor methods at each level to make it easier to first create a client and then to access elements that are further down the path.

Changed reviewed by Pavel B.

svn path=/trunk/; revision=308