This fix looks at trying to solve the problems in WADL-49 and WADL-50.
In WADL-49 we were presented with basically an invalid WADL; but the generation tool didn't
generate much in the way of useful debug information with regards to what the problems were,
indeed the process would fail with a null pointer exception because of the missing name
property on the top level method elements.

So I have contacted the original authors of the WADL, and presuming they fix the problems
it should be possible to generate a client, failing that the tool will at least give some
indication of why it isn't processing the file properly.

In WADL-50 we were looking at the same wadl, now hacked to be broadly correct with regards
to the "base" property. Unfortunately the original design assumed that the base property 
was an absolute URI, whereas the specification allows for the fact it could be relative path
(see xs:anyURI), so we have updated the generator to correctly create the right relative URI
if required. Also modified the Maven tool to take an array of URL in the form of the target
parameter so that it is possible to generate from a server rather than requiring a copy of
the files on disk.

svn path=/trunk/; revision=329