I have updated the Ant task so that it can take in custom class name mapping as part of its parameters:
 *   <wjc description="file.wadl" package="com.yahoo.search"
 *       autoSchemaPackage="true|false" target="gen-src">
 *     <customizations dir="." includes="binding.xjc"/>
 *     <produces dir="gen-src/com/yahoo/search" includes="*.java"/>
 *     <depends dir="." includes="schema.xsd"/>
 *     <customClassName uri="http://yahoo.com/rest" classname="YahooRest" >
 *   </wjc>

This is a speculative fix for a customer - there are no unit tests for the ant support
but the implementation appeared to be quite straight forward so I figured the risk was low.

There are two other files that are appearing in my diff appear to be left over from a previous
commit. Not sure what is going on here, but writing through for completeness. I suspect without
these the build would have been broken. 

svn path=/trunk/; revision=334