* Issue with methods with no response element  *
If there was a JAX-B parameter and but no response element, as in the case with for example a POST to create list, then
rather than void the method now return ClientReponse which is a bit more helpful.

* Issue with @XmlType wrapping *

When we were wrapping the input element we were using the return class rather than the input class. This would of course
fail, original test case was symmetrical so missing this issue.

* Proxy objects were immutable *

Made the proxy objects mutable, to be consistent with how it should work for JAX-RS 2.0

* No programmatic control of resource URI *

Resource URI is now provided as parameter for the static helper methods and for the constructors, this required
some simplification of where the URI parameters can be updated; but I think this is a better design all around.

svn path=/trunk/; revision=344