This is a interim commit that contains a partial implementation of a generator for JAX-RS 2.0 that is not quite complete yet. Most of the required refactoring is in place there are just a few implementation details missing for the generation to be complete. This is only for the current client API and doesn't generate for the Proxy API as yet.
The second part of this transaction is to do with improving the automated testing to include some basic functional testing
of the generated client. Until this tests assumed that that generated interface was working rather than trying to invoke the
client directly. This was done using FEST-Reflect to invoke the generated client. This is course is only test maven artifcats
but at the moment this is felt to be sufficient.

svn path=/trunk/; revision=346