* Empty webApplicationException message string
The default message for this exception with just a status was just an empty string, so this
code generates some boilerplate to fix this. (See JAX_RS_SPEC-312)

* Cannot process WADL with regex is template parameters (Oracle bug 16061180)

The code wasn't correctly extracting the parameter name if there was a regular expression in the 
template. This resulted in odd class names, and code that would fail to run because of missing 
parameters.

* Need minor update to work with latest version of Jersey 2.x

JerseyClient no longer returns a Configurable from a method configuation(),
so the boilerplate methods no longer work properly. Tweaked to code so
it now compiles.

* Update to surefire 2.8 / JUnit 4.x 

This is so in Netbeans 7.2.x you can run single tests in the debugger.

* Possibly a better fix for WADL-52

Make sure that the right Jersey elements are included in the MANIFEST.MF for the command
line version of the tool.

svn path=/trunk/; revision=365